refactor: res を :core:design-system に集約
