+++ core/design-system/src/main/res/drawable/icon_arrow_back.xml
