+++ core/design-system/src/main/res/drawable/icon_barcode_scanner.xml
