+++ core/design-system/src/main/res/drawable/icon_heart_filled.xml
