+++ core/design-system/src/main/res/drawable/icon_key.xml
