+++ core/design-system/src/main/res/drawable/icon_send.xml
