+++ core/design-system/src/main/res/values/strings.xml
@@ -1,5 +1,43 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+  <string name="my_book_detail_alt_add_my_book_to_favorites">Add your book to favorites</string>
+  <string name="my_book_detail_alt_archive_my_book">Archive your book</string>
+  <string name="my_book_detail_alt_back">Back to Mybrary</string>
+  <string name="my_book_detail_alt_background_image">Background image</string>
+  <string name="my_book_detail_alt_create_a_memo">Create a note</string>
+  <string name="my_book_detail_alt_make_my_book_private">Make your book private</string>
+  <string name="my_book_detail_alt_make_my_book_public">Make your book public</string>
+  <string name="my_book_detail_alt_remove_my_book_from_favorites">Remove your book from favorites</string>
+  <string name="my_book_detail_alt_save_a_memo">Save a note</string>
+  <string name="my_book_detail_message_my_book_added_to_your_favorites">Your book added to favorites</string>
+  <string name="my_book_detail_message_my_book_archived">Your book archived</string>
+  <string name="my_book_detail_message_my_book_is_now_private">Your book is now private</string>
+  <string name="my_book_detail_message_my_book_is_now_public">Your book is now public</string>
+  <string name="my_book_detail_message_my_book_removed_from_your_favorites">Your book removed from favorites</string>
+  <string name="my_book_detail_message_memo_created">Note created</string>
+  <string name="my_book_detail_message_memo_edited">Note edited</string>
+  <string name="my_book_detail_placeholder_end_page">End Page</string>
+  <string name="my_book_detail_placeholder_start_page">Start Page</string>
+  <string name="my_book_detail_placeholder_enter_memo">Note…</string>
+  <string name="my_book_detail_text_a">A</string>
+  <string name="my_book_detail_text_edited">%1$s • edited</string>
+  <string name="my_book_detail_text_p">p.%1$d</string>
+  <string name="my_book_detail_text_page_count">%1$d pages</string>
+  <string name="my_book_detail_text_page_count_and_publisher">%1$d pages｜ %2$s</string>
+  <string name="my_book_detail_text_pp">pp.%1$d~%2$d</string>
+  <string name="my_book_list_alt_search_for_books">Search for books</string>
+  <string name="my_book_list_text_a">A</string>
+  <string name="my_book_list_text_all_my_books">All Books</string>
+  <string name="my_book_list_text_favorite_my_books">Favorite Books</string>
+  <string name="my_book_list_text_other_my_books">Other Books</string>
+  <string name="search_book_alt_scan_book_barcode">Scan book barcode</string>
+  <string name="search_book_alt_search_for_books">Search for books</string>
+  <string name="search_book_placeholder_enter_search_keywords">Search…</string>
+  <string name="search_book_text_a">A</string>
+  <string name="search_book_text_add">Add</string>
+  <string name="search_book_text_a_a">A\nA</string>
+  <string name="search_book_text_cancel">Cancel</string>
+  <string name="search_book_text_would_you_like_to_add_to_mybrary">Would you like to add to Mybrary?</string>
   <string name="setting_list_text_about">About</string>
   <string name="setting_list_text_account">Account</string>
   <string name="setting_list_text_cancel">Cancel</string>
@@ -15,8 +53,37 @@
   <string name="setting_list_text_terms_of_use">Terms of Use</string>
   <string name="setting_list_text_version">Version</string>
   <string name="setting_list_text_would_you_like_to_logout">Would you like to logout?</string>
+  <string name="sign_in_alt_sign_in_with_google">Sign in with Google</string>
+  <string name="sign_in_text_create_an_account">"Create an account "</string>
+  <string name="sign_in_text_here">here</string>
+  <string name="sign_in_text_sign_in_with_google">Sign in with Google</string>
+  <string name="sign_up_alt_continue_as_anonymous">Continue as anonymous</string>
+  <string name="sign_up_alt_send_otp">Send One-Time Password</string>
+  <string name="sign_up_alt_sign_up_with_google">Sign up with Google</string>
+  <string name="sign_up_alt_your_email">Your Email</string>
+  <string name="sign_up_placeholder_enter_your_email">Your Email</string>
+  <string name="sign_up_text_already_have_an_account_login">"Already have an account? Login "</string>
+  <string name="sign_up_text_continue_as_anonymous">Continue as anonymous</string>
+  <string name="sign_up_text_here">here</string>
+  <string name="sign_up_text_send_otp">Send One-Time Password</string>
+  <string name="sign_up_text_sign_up_with_google">Sign up with Google</string>
+  <string name="ui_alt_book_cover">Book cover</string>
   <string name="ui_alt_my_book">Go to Mybrary destination</string>
   <string name="ui_alt_setting">Go to Settings destination</string>
+  <string name="ui_text_and">" and "</string>
+  <string name="ui_text_apply">" apply."</string>
   <string name="ui_text_my_book">Mybrary</string>
+  <string name="ui_text_or">or</string>
+  <string name="ui_text_privacy_policy">Privacy Policy</string>
   <string name="ui_text_setting">Settings</string>
+  <string name="ui_text_terms_of_use">Terms of Use</string>
+  <string name="verify_otp_alt_otp">One-Time Password</string>
+  <string name="verify_otp_alt_back">Back to send One-Time Password</string>
+  <string name="verify_otp_placeholder_enter_otp">123456</string>
+  <string name="verify_otp_text_enter_otp_sent_to">Enter the One-Time Password sent to %1$s.</string>
+  <string name="verify_otp_text_it_may_take_a_few_minutes">It may take a few minutes to receive the one-time password.</string>
+  <string name="verify_otp_text_login">Login</string>
+  <string name="verify_otp_text_resend_otp">Resend One-Time Password</string>
+  <string name="verify_otp_text_sign_up">Sign up</string>
+  <string name="verify_otp_text_verify_otp">Verify One-Time Password</string>
 </resources>
