--- core/ui/src/main/res/values-ja/strings.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-  <string name="my_book_detail_alt_add_my_book_to_favorites">本をお気に入りに追加する</string>
-  <string name="my_book_detail_alt_archive_my_book">本をアーカイブする</string>
-  <string name="my_book_detail_alt_back">Mybrary に戻る</string>
-  <string name="my_book_detail_alt_background_image">背景画像</string>
-  <string name="my_book_detail_alt_create_a_memo">メモを作成する</string>
-  <string name="my_book_detail_alt_make_my_book_private">本を非公開にする</string>
-  <string name="my_book_detail_alt_make_my_book_public">本を公開する</string>
-  <string name="my_book_detail_alt_remove_my_book_from_favorites">本をお気に入りから削除する</string>
-  <string name="my_book_detail_alt_save_a_memo">メモを保存する</string>
-  <string name="my_book_detail_message_my_book_added_to_your_favorites">お気に入りに追加しました</string>
-  <string name="my_book_detail_message_my_book_archived">アーカイブしました</string>
-  <string name="my_book_detail_message_my_book_is_now_private">公開しました</string>
-  <string name="my_book_detail_message_my_book_is_now_public">非公開にしました</string>
-  <string name="my_book_detail_message_my_book_removed_from_your_favorites">お気に入りから削除しました</string>
-  <string name="my_book_detail_message_memo_created">メモを作成しました</string>
-  <string name="my_book_detail_message_memo_edited">メモを編集しました</string>
-  <string name="my_book_detail_placeholder_end_page">終了ページ</string>
-  <string name="my_book_detail_placeholder_start_page">開始ページ</string>
-  <string name="my_book_detail_placeholder_enter_memo">メモを入力…</string>
-  <string name="my_book_detail_text_a">あ</string>
-  <string name="my_book_detail_text_edited">%1$s（編集済み）</string>
-  <string name="my_book_detail_text_p">p.%1$d</string>
-  <string name="my_book_detail_text_page_count">%1$dページ</string>
-  <string name="my_book_detail_text_page_count_and_publisher">%1$dページ｜ %2$s</string>
-  <string name="my_book_detail_text_pp">pp.%1$d~%2$d</string>
-  <string name="my_book_list_alt_search_for_books">本を検索する</string>
-  <string name="my_book_list_text_a">あ</string>
-  <string name="my_book_list_text_all_my_books">すべての本</string>
-  <string name="my_book_list_text_favorite_my_books">お気に入りの本</string>
-  <string name="my_book_list_text_other_my_books">その他の本</string>
-  <string name="search_book_alt_scan_book_barcode">バーコードを読み取る</string>
-  <string name="search_book_alt_search_for_books">本を検索する</string>
-  <string name="search_book_placeholder_enter_search_keywords">検索ワードを入力…</string>
-  <string name="search_book_text_a">あ</string>
-  <string name="search_book_text_add">追加する</string>
-  <string name="search_book_text_a_a">あ\nあ</string>
-  <string name="search_book_text_cancel">追加しない</string>
-  <string name="search_book_text_would_you_like_to_add_to_mybrary">Mybrary に本を追加しますか？</string>
-  <string name="sign_in_alt_sign_in_with_google">Google アカウントでログイン</string>
-  <string name="sign_in_text_create_an_account">アカウント作成は</string>
-  <string name="sign_in_text_here">こちら</string>
-  <string name="sign_in_text_sign_in_with_google">Google アカウントでログイン</string>
-  <string name="sign_up_alt_continue_as_anonymous">ゲストで始めてみる</string>
-  <string name="sign_up_alt_send_otp">ワンタイムパスワードを送信する</string>
-  <string name="sign_up_alt_sign_up_with_google">Google アカウントで始める</string>
-  <string name="sign_up_alt_your_email">メールアドレス</string>
-  <string name="sign_up_placeholder_enter_your_email">メールアドレス</string>
-  <string name="sign_up_text_already_have_an_account_login">ログインは</string>
-  <string name="sign_up_text_continue_as_anonymous">ゲストで始めてみる</string>
-  <string name="sign_up_text_here">こちら</string>
-  <string name="sign_up_text_send_otp">ワンタイムパスワードを送信する</string>
-  <string name="sign_up_text_sign_up_with_google">Google アカウントで始める</string>
-  <string name="ui_alt_book_cover">本の表紙</string>
-  <string name="ui_text_and">と</string>
-  <string name="ui_text_apply">が適用されます。</string>
-  <string name="ui_text_or">または</string>
-  <string name="ui_text_privacy_policy">プライバシーポリシー</string>
-  <string name="ui_text_terms_of_use">利用規約</string>
-  <string name="verify_otp_alt_otp">ワンタイムパスワード</string>
-  <string name="verify_otp_alt_back">ワンタイムパスワードの送信に戻る</string>
-  <string name="verify_otp_placeholder_enter_otp">123456</string>
-  <string name="verify_otp_text_enter_otp_sent_to">%1$sに送信されたワンタイムパスワードを入力してください。</string>
-  <string name="verify_otp_text_it_may_take_a_few_minutes">ワンタイムパスワードが届くまで数分かかる場合があります。</string>
-  <string name="verify_otp_text_login">ログイン</string>
-  <string name="verify_otp_text_resend_otp">もう一度ワンタイムパスワードを送信</string>
-  <string name="verify_otp_text_sign_up">サインアップ</string>
-  <string name="verify_otp_text_verify_otp">ワンタイムパスワードの認証</string>
-</resources>
