+++ feature/sign-in/src/main/java/app/kaito_dogi/mybrary/feature/signin/component/NavigateToSignUpSection.kt
@@ -13,8 +13,8 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.withLink
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
+import app.kaito_dogi.mybrary.core.designsystem.R
 import app.kaito_dogi.mybrary.core.designsystem.theme.MybraryTheme
-import app.kaito_dogi.mybrary.core.ui.R
 
 @Composable
 internal fun NavigateToSignUpSection(
