+++ feature/verify-otp/src/main/java/app/kaito_dogi/mybrary/feature/verifyotp/component/VerifyOtpTopAppBar.kt
@@ -9,8 +9,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
+import app.kaito_dogi.mybrary.core.designsystem.R
 import app.kaito_dogi.mybrary.core.designsystem.theme.MybraryTheme
-import app.kaito_dogi.mybrary.core.ui.R
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
